pyproject.toml
[project]
name = "ecids-test"
version = "0.1.0"
description = "ECIDS Test"
authors = [
    {name = "CSW191",email = "joshua.parep@mt.gov"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9, <3.14"
dependencies = [
    "dbt-snowflake (>=1.10.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "dagster (>=1.11.4,<2.0.0)",
    "dagster-dbt (>=0.27.4,<0.28.0)",
    "dagster-webserver (>=1.11.4,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"

